+++ CHANGELOG.md
@@ -24,6 +24,10 @@
 * Fix migration task bug that causes migraiton to fail on old xenon versions.
   https://www.pivotaltracker.com/story/show/150185195
 
+* Added support for request logging that logs all inbound requests to a Service
+  Host. Logging can be enabled or disabled during host start-up by calling
+  ServiceHost.setRequestLoggingInfo OR by making a PATCH to ServiceHostManagementService.
+
 ## 1.5.4
 
 * Updates to migration task service to assert zero documentOwner mismatches